requirements.txt
annotated-types==0.7.0
asttokens==3.0.1
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
decorator==5.2.1
eval_type_backport==0.3.1
exceptiongroup==1.3.1
executing==2.2.1
filelock==3.19.1
fonttools==4.60.2
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
idna==3.11
importlib_resources==6.5.2
ipdb==0.13.13
ipython==8.18.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.4
matplotlib-inline==0.2.1
mpmath==1.3.0
networkx==3.2.1
ninja==1.13.0
numpy==1.26.4
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
protobuf==6.33.2
psutil==7.2.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
scikit-learn==1.6.1
scipy==1.13.1
sentry-sdk==2.48.0
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
sympy==1.14.0
threadpoolctl==3.6.0
tomli==2.3.0
torch==2.1.2+cu121
torch-cluster==1.6.3+pt21cu121
torch-scatter==2.1.2+pt21cu121
torch-sparse==0.6.18+pt21cu121
torch-spline-conv==1.2.2+pt21cu121
torch_geometric==2.4.0
tqdm==4.67.1
traitlets==5.14.3
triton==2.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
wandb==0.23.1
wcwidth==0.2.14
zipp==3.23.0